Add files via upload
Added in the following scenes and their scripts:

Global - holds the variables needed for scoring and it's functions
Game - visualizes the notes on screen when called from Conductor node
Conductor - main engine of the game. tracks and creates the notes based on the song's beat and song position
End - temporary end screen with back to menu and play again buttons
Note - main track for the notes to flow on. currently configurated in top/down orientation. needs to be adjusted and tweaked for right/left config.